requirements.txt
langchain_core
streamlit
langchain_groq
langchain
groq
base64
logging
pyyaml
